Use float and finish normal interpolation

+++ deps/vecmath/include/Matrix2f.h
@@ -10,9 +10,9 @@ class Matrix2f
 {
 public:
 	// Fill a 2x2 matrix with "fill", default to 0.
-	Matrix2f(double fill = 0.f);
-	Matrix2f(double m00, double m01,
-			 double m10, double m11);
+	Matrix2f(float fill = 0.f);
+	Matrix2f(float m00, float m01,
+			 float m10, float m11);
 
 	// setColumns = true ==> sets the columns of the matrix to be [v0 v1]
 	// otherwise, sets the rows
@@ -22,8 +22,8 @@ public:
 	Matrix2f &operator=(const Matrix2f &rm); // assignment operator
 	// no destructor necessary
 
-	const double &operator()(int i, int j) const;
-	double &operator()(int i, int j);
+	const float &operator()(int i, int j) const;
+	float &operator()(int i, int j);
 
 	Vector2f getRow(int i) const;
 	void setRow(int i, const Vector2f &v);
@@ -31,30 +31,30 @@ public:
 	Vector2f getCol(int j) const;
 	void setCol(int j, const Vector2f &v);
 
-	double determinant();
-	Matrix2f inverse(bool *pbIsSingular = NULL, double epsilon = 0.f);
+	float determinant();
+	Matrix2f inverse(bool *pbIsSingular = NULL, float epsilon = 0.f);
 
 	void transpose();
 	Matrix2f transposed() const;
 
 	// ---- Utility ----
-	operator double *(); // automatic type conversion for GL
+	operator float *(); // automatic type conversion for GL
 	void print();
 
-	static double determinant2x2(double m00, double m01,
-								 double m10, double m11);
+	static float determinant2x2(float m00, float m01,
+								 float m10, float m11);
 
 	static Matrix2f ones();
 	static Matrix2f identity();
-	static Matrix2f rotation(double degrees);
+	static Matrix2f rotation(float degrees);
 
 private:
-	double m_elements[4];
+	float m_elements[4];
 };
 
 // Scalar-Matrix multiplication
-Matrix2f operator*(double f, const Matrix2f &m);
-Matrix2f operator*(const Matrix2f &m, double f);
+Matrix2f operator*(float f, const Matrix2f &m);
+Matrix2f operator*(const Matrix2f &m, float f);
 
 // Matrix-Vector multiplication
 // 2x2 * 2x1 ==> 2x1
