+++ deps/vecmath/src/Vector3f.cpp
@@ -21,28 +21,28 @@ const Vector3f Vector3f::RIGHT = Vector3f(1, 0, 0);
 // static
 const Vector3f Vector3f::FORWARD = Vector3f(0, 0, -1);
 
-Vector3f::Vector3f(double f)
+Vector3f::Vector3f(float f)
 {
 	m_elements[0] = f;
 	m_elements[1] = f;
 	m_elements[2] = f;
 }
 
-Vector3f::Vector3f(double x, double y, double z)
+Vector3f::Vector3f(float x, float y, float z)
 {
 	m_elements[0] = x;
 	m_elements[1] = y;
 	m_elements[2] = z;
 }
 
-Vector3f::Vector3f(const Vector2f &xy, double z)
+Vector3f::Vector3f(const Vector2f &xy, float z)
 {
 	m_elements[0] = xy.x();
 	m_elements[1] = xy.y();
 	m_elements[2] = z;
 }
 
-Vector3f::Vector3f(double x, const Vector2f &yz)
+Vector3f::Vector3f(float x, const Vector2f &yz)
 {
 	m_elements[0] = x;
 	m_elements[1] = yz.x();
@@ -67,42 +67,42 @@ Vector3f &Vector3f::operator=(const Vector3f &rv)
 	return *this;
 }
 
-const double &Vector3f::operator[](int i) const
+const float &Vector3f::operator[](int i) const
 {
 	return m_elements[i];
 }
 
-double &Vector3f::operator[](int i)
+float &Vector3f::operator[](int i)
 {
 	return m_elements[i];
 }
 
-double &Vector3f::x()
+float &Vector3f::x()
 {
 	return m_elements[0];
 }
 
-double &Vector3f::y()
+float &Vector3f::y()
 {
 	return m_elements[1];
 }
 
-double &Vector3f::z()
+float &Vector3f::z()
 {
 	return m_elements[2];
 }
 
-double Vector3f::x() const
+float Vector3f::x() const
 {
 	return m_elements[0];
 }
 
-double Vector3f::y() const
+float Vector3f::y() const
 {
 	return m_elements[1];
 }
 
-double Vector3f::z() const
+float Vector3f::z() const
 {
 	return m_elements[2];
 }
@@ -137,12 +137,12 @@ Vector3f Vector3f::zxy() const
 	return Vector3f(m_elements[2], m_elements[0], m_elements[1]);
 }
 
-double Vector3f::length() const
+float Vector3f::length() const
 {
 	return sqrt(m_elements[0] * m_elements[0] + m_elements[1] * m_elements[1] + m_elements[2] * m_elements[2]);
 }
 
-double Vector3f::squaredLength() const
+float Vector3f::squaredLength() const
 {
 	return (
 		m_elements[0] * m_elements[0] +
@@ -152,7 +152,7 @@ double Vector3f::squaredLength() const
 
 void Vector3f::normalize()
 {
-	double norm = length();
+	float norm = length();
 	m_elements[0] /= norm;
 	m_elements[1] /= norm;
 	m_elements[2] /= norm;
@@ -160,7 +160,7 @@ void Vector3f::normalize()
 
 Vector3f Vector3f::normalized() const
 {
-	double norm = length();
+	float norm = length();
 	return Vector3f(
 		m_elements[0] / norm,
 		m_elements[1] / norm,
@@ -181,12 +181,12 @@ void Vector3f::negate()
 	m_elements[2] = -m_elements[2];
 }
 
-Vector3f::operator const double *() const
+Vector3f::operator const float *() const
 {
 	return m_elements;
 }
 
-Vector3f::operator double *()
+Vector3f::operator float *()
 {
 	return m_elements;
 }
@@ -213,7 +213,7 @@ Vector3f &Vector3f::operator-=(const Vector3f &v)
 	return *this;
 }
 
-Vector3f &Vector3f::operator*=(double f)
+Vector3f &Vector3f::operator*=(float f)
 {
 	m_elements[0] *= f;
 	m_elements[1] *= f;
@@ -222,7 +222,7 @@ Vector3f &Vector3f::operator*=(double f)
 }
 
 // static
-double Vector3f::dot(const Vector3f &v0, const Vector3f &v1)
+float Vector3f::dot(const Vector3f &v0, const Vector3f &v1)
 {
 	return v0[0] * v1[0] + v0[1] * v1[1] + v0[2] * v1[2];
 }
@@ -237,13 +237,13 @@ Vector3f Vector3f::cross(const Vector3f &v0, const Vector3f &v1)
 }
 
 // static
-Vector3f Vector3f::lerp(const Vector3f &v0, const Vector3f &v1, double alpha)
+Vector3f Vector3f::lerp(const Vector3f &v0, const Vector3f &v1, float alpha)
 {
 	return alpha * (v1 - v0) + v0;
 }
 
 // static
-Vector3f Vector3f::cubicInterpolate(const Vector3f &p0, const Vector3f &p1, const Vector3f &p2, const Vector3f &p3, double t)
+Vector3f Vector3f::cubicInterpolate(const Vector3f &p0, const Vector3f &p1, const Vector3f &p2, const Vector3f &p3, float t)
 {
 	// geometric construction:
 	//            t
@@ -288,17 +288,17 @@ Vector3f operator-(const Vector3f &v)
 	return Vector3f(-v[0], -v[1], -v[2]);
 }
 
-Vector3f operator*(double f, const Vector3f &v)
+Vector3f operator*(float f, const Vector3f &v)
 {
 	return Vector3f(v[0] * f, v[1] * f, v[2] * f);
 }
 
-Vector3f operator*(const Vector3f &v, double f)
+Vector3f operator*(const Vector3f &v, float f)
 {
 	return Vector3f(v[0] * f, v[1] * f, v[2] * f);
 }
 
-Vector3f operator/(const Vector3f &v, double f)
+Vector3f operator/(const Vector3f &v, float f)
 {
 	return Vector3f(v[0] / f, v[1] / f, v[2] / f);
 }
