+++ include/group.hpp
@@ -26,7 +26,7 @@ public:
             delete object_p;
     }
 
-    bool intersect(const Ray &r, Hit &h, double t_min) override
+    bool intersect(const Ray &r, Hit &h, float t_min) override
     {
         bool intersected = false;
         for (auto object_p : objects)
